+++ backend/controllers/students/DefaultController.php
@@ -9,6 +9,7 @@ use backend\models\Model;
 use common\models\education\EducationCat;
 use common\models\education\EducationProgrammLevel;
 use common\models\history\StudyplanHistory;
+use common\models\sigur\UsersCard;
 use common\models\students\StudentDependence;
 use common\models\studyplan\search\StudyplanSearch;
 use common\models\studyplan\Studyplan;
@@ -39,17 +40,19 @@ class DefaultController extends MainController
 
         $user = new User();
         $userCommon = new UserCommon();
+        $userCard = new UsersCard();
        // $userCommon->scenario = UserCommon::SCENARIO_NEW;
         $model = new $this->modelClass;
         $modelsDependence = [new StudentDependence(['scenario' => StudentDependence::SCENARIO_PARENT])];
 
-        if ($userCommon->load(Yii::$app->request->post()) && $model->load(Yii::$app->request->post())) {
+        if ($userCommon->load(Yii::$app->request->post()) && $userCard->load(Yii::$app->request->post()) && $model->load(Yii::$app->request->post())) {
 
             $modelsDependence = Model::createMultiple(StudentDependence::class);
             Model::loadMultiple($modelsDependence, Yii::$app->request->post());
 
             // validate all models
             $valid = $userCommon->validate();
+            $valid = $userCard->validate() && $valid;
             $valid = $model->validate() && $valid;
             $valid = Model::validateMultiple($modelsDependence) && $valid;
             //$valid = true;
@@ -69,13 +72,16 @@ class DefaultController extends MainController
                         $userCommon->user_category = UserCommon::USER_CATEGORY_STUDENTS;
                         $userCommon->user_id = $user->id;
                         if ($flag = $userCommon->save(false)) {
-                            $model->user_common_id = $userCommon->id;
-                            if ($flag = $model->save(false)) {
-                                foreach ($modelsDependence as $modelDependence) {
-                                    $modelDependence->student_id = $model->id;
-                                    if (!($flag = $modelDependence->save(false))) {
-                                        $transaction->rollBack();
-                                        break;
+                            $userCard->user_common_id = $userCommon->id;
+                            if ($flag = $userCard->save(false)) {
+                                $model->user_common_id = $userCommon->id;
+                                if ($flag = $model->save(false)) {
+                                    foreach ($modelsDependence as $modelDependence) {
+                                        $modelDependence->student_id = $model->id;
+                                        if (!($flag = $modelDependence->save(false))) {
+                                            $transaction->rollBack();
+                                            break;
+                                        }
                                     }
                                 }
                             }
@@ -94,6 +100,7 @@ class DefaultController extends MainController
 
         return $this->renderIsAjax('create', [
             'userCommon' => $userCommon,
+            'userCard' => $userCard,
             'model' => $model,
             'modelsDependence' => (empty($modelsDependence)) ? [new StudentDependence] : $modelsDependence,
             'readonly' => false
@@ -113,6 +120,7 @@ class DefaultController extends MainController
 
         $model = $this->findModel($id);
         $userCommon = UserCommon::findOne(['id' => $model->user_common_id, 'user_category' => UserCommon::USER_CATEGORY_STUDENTS]);
+        $userCard = UsersCard::findOne(['user_common_id' => $model->user_common_id]) ?: new UsersCard();
        // $userCommon->scenario = UserCommon::SCENARIO_UPDATE;
 
         if (!isset($model, $userCommon)) {
@@ -123,7 +131,7 @@ class DefaultController extends MainController
         foreach ($modelsDependence as $m) {
             $m->scenario = StudentDependence::SCENARIO_PARENT;
         }
-        if ($userCommon->load(Yii::$app->request->post()) && $model->load(Yii::$app->request->post())) {
+        if ($userCommon->load(Yii::$app->request->post()) && $userCard->load(Yii::$app->request->post()) && $model->load(Yii::$app->request->post())) {
 
             $oldIDs = ArrayHelper::map($modelsDependence, 'id', 'id');
             $modelsDependence = Model::createMultiple(StudentDependence::class, $modelsDependence);
@@ -132,6 +140,7 @@ class DefaultController extends MainController
 
             // validate all models
             $valid = $userCommon->validate();
+            // $valid = $userCard->validate() && $valid;
             $valid = $model->validate() && $valid;
             $valid = Model::validateMultiple($modelsDependence) && $valid;
 
@@ -139,15 +148,18 @@ class DefaultController extends MainController
                 $transaction = \Yii::$app->db->beginTransaction();
                 try {
                     if ($flag = $userCommon->save(false)) {
-                        if ($flag = $model->save(false)) {
-                            if (!empty($deletedIDs)) {
-                                StudentDependence::deleteAll(['id' => $deletedIDs]);
-                            }
-                            foreach ($modelsDependence as $modelDependence) {
-                                $modelDependence->student_id = $model->id;
-                                if (!($flag = $modelDependence->save(false))) {
-                                    $transaction->rollBack();
-                                    break;
+                        $userCard->user_common_id = $userCommon->id;
+                        if ($flag && $flag = $userCard->save(false)) {
+                            if ($flag = $model->save(false)) {
+                                if (!empty($deletedIDs)) {
+                                    StudentDependence::deleteAll(['id' => $deletedIDs]);
+                                }
+                                foreach ($modelsDependence as $modelDependence) {
+                                    $modelDependence->student_id = $model->id;
+                                    if (!($flag = $modelDependence->save(false))) {
+                                        $transaction->rollBack();
+                                        break;
+                                    }
                                 }
                             }
                         }
@@ -164,6 +176,7 @@ class DefaultController extends MainController
 
         return $this->render('update', [
             'userCommon' => $userCommon,
+            'userCard' => $userCard,
             'model' => $model,
             'modelsDependence' => (empty($modelsDependence)) ? [new StudentDependence] : $modelsDependence,
             'readonly' => $readonly
